software_engineering/01/pyproject.toml
[project]
name = "swe_hw_1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mock>=5.2.0",
    "opyoid>=3.0.3",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-ulid>=3.1.0",
    "questionary>=2.1.1",
    "termcolor>=3.1.0",
    "typing-extensions>=4.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/swe_hw_1"]
